+++ src/tools.cpp
@@ -45,37 +45,37 @@ VectorXd Tools::CalculateRMSE(const vector<VectorXd> &estimations,
 
 MatrixXd Tools::CalculateJacobian(const VectorXd& x_state) {
 	MatrixXd Hj(3, 4);
-	float minThresh = 0.0001;
+	double minThresh = 0.0001;
 	//recover state parameters
-	float px = (fabs(x_state(0)) < 0.001) ? 0.001 : x_state(0);
-	float py = (fabs(x_state(1)) < 0.001) ? 0.001 : x_state(1);
-	float vx = x_state(2);
-	float vy = x_state(3);
+	double px = (fabs(x_state(0)) < 0.001) ? 0.001 : x_state(0);
+	double py = (fabs(x_state(1)) < 0.001) ? 0.001 : x_state(1);
+	double vx = x_state(2);
+	double vy = x_state(3);
 
-	float p2 = px*px + py*py;
+	double p2 = px*px + py*py;
 	if (p2 <= minThresh)
 	{
 		cout << "Warning - Tools::CalculateJacobian - low px/py values" << endl;
 		p2 = minThresh;
 	}
 
-	float den = sqrt(p2);	
-	float dpdpx = px / den;
-	float dpdpy = py / den;
-	float dpdvx = 0.0;
-	float dpdvy = 0.0;
+	double den = sqrt(p2);
+	double dpdpx = px / den;
+	double dpdpy = py / den;
+	double dpdvx = 0.0;
+	double dpdvy = 0.0;
 
 	den = p2;
-	float dphidpx = -py / den;
-	float dphidpy = px / den;
-	float dphidvx = 0.0;
-	float dphidvy = 0.0;
+	double dphidpx = -py / den;
+	double dphidpy = px / den;
+	double dphidvx = 0.0;
+	double dphidvy = 0.0;
 
 	den = pow((pow(px, 2) + pow(py, 2)), 3.0 / 2.0);
-	float dprdpx = (vx*py - vy*px)*py / den;
-	float dprdpy = (vy*px - vx*py)*px / den;
-	float dprdvx = dpdpx;
-	float dprdvy = dpdpy;
+	double dprdpx = (vx*py - vy*px)*py / den;
+	double dprdpy = (vy*px - vx*py)*px / den;
+	double dprdvx = dpdpx;
+	double dprdvy = dpdpy;
 
 	//compute the Jacobian matrix
 	Hj << dpdpx, dpdpy, dpdvx, dpdvy,
